mobile_automation/requirements.txt
Appium-Python-Client==4.0.0
pytest==8.3.2
pytest-timeout==2.3.1